refactor(config): split utils
