+++ src/config/src/utils/index.ts
@@ -1,0 +1,5 @@
+export * as compositeRules from "./compositeRules";
+export * as rules from "./rules";
+export * from "./defaultValues";
+export * from "./severity";
+export * from "./templates";
